#334 move geometry loader into geometry dir
